# swagger/ft.yaml
openapi: 3.0.0
info:
  title: ftx
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
...
  '/api/user/search':
    get:
      tags: [ users ]
      summary: Searching for users - admin only
      operationId: get-api-user-search
      description: Searching for users
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "q"
          in: 'query'
          description: "Query string for searching"
          schema:
            type: "string"
        - name: "email"
          in: 'query'
          description: "User's email"
          schema:
            type: "string"
        - name: "name"
          in: 'query'
          description: "User's name"
          required: false
          schema:
            type: "string"
        - name: "surname"
          in: 'query'
          description: "User's surname"
          schema:
            type: "string"
        - name: "role"
          in: 'query'
          description: "User's role"
          schema:
            type: "string"
        - name: "id"
          in: 'query'
          description: "User's id"
          schema:
            type: "string"
        - name: "orgname"
          in: 'query'
          description: "The name of the organization to which the user belongs"
          schema:
            type: "string"
            default: null
        - name: "orgid"
          in: 'query'
          description: "The id of the organization to which the user belongs"
          schema:
            type: "string"
        - name: "groupname"
          in: 'query'
          description: "The name of the group to which the user belongs"
          schema:
            type: "string"
        - name: "groupid"
          in: 'query'
          description: "The id of the group to which the user belongs"
          schema:
            type: "string"
        - name: "lang"
          in: 'query'
          description: "User's language in ISO 639-1 format"
          schema:
            type: "string"
            items:
              type: "string"
              enum:
                - "en"
                - "fr"
        - name: "sub"
          in: 'query'
          description: "Flag that expands user nested objects"
          schema:
            type: "boolean"
        - name: 'pageSize'
          in: 'query'
          description: "Maximum number of items per page after the search"
          schema:
            type: "string"
      responses:
        '200':
          description: Searching for users
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/AggregatedUserProfile'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileId'
...
  /api/preuser:
    put:
      tags: [ preuser ]
      summary: Create new preuser
      operationId: put-preuser
      description: Create new preuser
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "notify"
          in: 'query'
          description: "Send notify for user after pre-user has been created if true"
          schema:
            type: "boolean"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userData
              properties:
                userData:
                  type: object
                  description: all preuser data
                  properties:
                    name:
                      type: string
                    surname:
                      type: string
                    role:
                      type: string
                      enum:
                        - founder
                        - investor
                        - admin
                    email:
                      type: string
                    avatarUrl:
                      type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/PreuserCreation'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: []
  '/api/preuser/{preuserId}':
    get:
      tags: [ preuser ]
      summary: Fetches preuser by id
      operationId: get-preuser
      description: Retrieves existing preuser
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "preuserId"
          in: 'path'
          description: "preuserId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/PreuserCreation'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: []
    post:
      tags: [ preuser ]
      summary: Updates preuser by id
      operationId: post-preuser
      description: Updates existing preuser
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "preuserId"
          in: 'path'
          description: "preuserId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreuserCreation'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreuserEditable'
      security:
        - API Key: []
    delete:
      tags: [ preuser ]
      summary: Deletes preuser by id
      operationId: delete-preuser
      description: Deletes existing preuser
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "preuserId"
          in: 'path'
          description: "preuserId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/PreuserCreation'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: []
  '/api/preuser/search':
    get:
      tags: [ preuser ]
      summary: Searching for preusers
      operationId: get-preuser-search
      description: Searching for preusers
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "q"
          in: 'query'
          description: "Query string for searching (serarches by name, surname, email, case insansative)"
          schema:
            type: "string"
        - name: "email"
          in: 'query'
          description: "Preuser's email"
          schema:
            type: "string"
        - name: "name"
          in: 'query'
          description: "Preuser's name"
          required: false
          schema:
            type: "string"
        - name: "surname"
          in: 'query'
          description: "Preuser's surname"
          schema:
            type: "string"
        - name: "role"
          in: 'query'
          description: "Preuser's role"
          schema:
            type: "string"
        - name: "inviteCodeSent"
          in: 'query'
          description: "Searches only preusers to whom the invite code was sent, if true"
          schema:
            type: "boolean"
        - name: 'pageSize'
          in: 'query'
          description: "Maximum number of items per page after the search"
          schema:
            type: "string"
      responses:
        '200':
          description: Searching for preusers
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/PreuserCreation'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
  '/api/preuser/register':
    get:
      tags: [ preuser ]
      summary: User registration
      operationId: get-preuser-register
      description: User registration by preuser invitation code
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "code"
          in: 'query'
          description: "Authorization code that user has been received on email, code format 'xxx-xxx'"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    example: { preUserId: 'preUserId', firstName: 'user first name' }
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: []
  '/api/preuser/invite/{preuserId}':
    get:
      tags: [ preuser ]
      summary: Send invite code to preuser
      operationId: get-preuser-invite
      description: Send invite code to preuser with preuser id
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "preuserId"
          in: 'path'
          description: "preuserId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/PreuserCreation'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: []
  '/api/preuser/invite/list':
    get:
      tags: [ preuser ]
      summary: Send invite code to a list of preusers
      operationId: get-preuser-invite-list
      description: Send invite code to a list of preusers
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/PreusersIdListSuccesfullySent'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreusersInviteList'
        description: Subset of UserProfile which regular non-privileged user is allowed to update
      security:
        - API Key: []
...
components:
...
  schemas:
...
    UserProfileId:
      description: ''
      type: object
      x-examples:
        UserProfileId example:
          id: string
      title: UserProfileId
      properties:
        id:
          type: string
          minLength: 1
          description: |
            Unique user ID
          example: 5fc52fbcf779cd786fabfff0
      required:
        - id
...
    PreuserCreation:
      description: 'preuser'
      type: object
      title: PreuserEditable
      parameters:
        userData:
          type: object
          description: object with all preuser data
          properties:
            userData:
              type: object
              required: true
              properties:
                email:
                  type: string
                role:
                  type: string
                name:
                  type: string
                surname:
                  type: string
                avatarUrl:
                  type: string
    PreusersInviteList:
      description: 'Array of preuser ids'
      type: array
      example: [
          "5fd61c19631e2d86c5ae9cf0",
          "6fd61c19631e2d86c5ae9cf1",
          "7fd61c19631e2d86c5ae9cf2",
      ]
    PreusersIdListSuccesfullySent:
      description: 'Array of preuser ids that invite code has been succesfully sent'
      type: array
      example: [
          "5fd61c19631e2d86c5ae9cf0",
          "6fd61c19631e2d86c5ae9cf1",
          "7fd61c19631e2d86c5ae9cf2",
      ]
    PreuserEditable:
      description: 'Edit preuser'
      type: object
      x-examples:
        example-1:
          email: string
          name: string
          surname: string
          avatarUrl: string
          role: string
      title: PreuserEditable
      properties:
        email:
          type: string
          minLength: 1
          description: preuser email. Should be unique across the preuser
          example: test@domain.com
        name:
          type: string
          minLength: 1
          description: Preuser's first name
          example: BoJack
        surname:
          type: string
          minLength: 1
          description: Preuser's last name
          example: Horseman
        role:
          type: string
          description: Preuser's role
          enum:
            - founder
            - admin
            - investor
    UserProfileUserEditable:
      description: ''
      type: object
      x-examples:
        example-1:
          email: string
          name: string
          surname: string
          avatarUrl: string
          bio: string
          theme: string
          watchedVideos: 0
      title: UserProfileUserEditable
      properties:
        email:
          type: string
          minLength: 1
          description: User email. Should be unique across the users
          example: test@domain.com
        name:
          type: string
          minLength: 1
          description: User first name
          example: BoJack
        surname:
          type: string
          minLength: 1
          description: User last name
          example: Horseman
        avatarUrl:
          type: string
          minLength: 1
          description: User's avatar URL
          example: 'https://i.pravatar.cc/300'
        bio:
          type: string
          minLength: 1
          description: User's bio
          example: User's bio goes here
        theme:
          type: string
          minLength: 1
          description: User's client application theme
          enum:
            - light
            - dark
    UserProfileAdminEditable:
      description: ''
      type: object
      x-examples:
        example-1:
          createdAt: string
          verified: true
          deleted: true
          paid: true
          role: string
          hasMarketingConsent: true
          marketingConsentTimestamp: string
          watchedVideos: 0
      title: UserProfileAdminEditable
      properties:
        createdAt:
          type: string
          minLength: 1
          description: Date and time profile was created
          example: '2020-11-30T17:45:32.193Z'
        verified:
          type: boolean
          description: User verification status
        deleted:
          type: boolean
          description: True indicates that user profile has been disabled or deleted
        paid:
          type: boolean
          description: True indicates that user profile is of paying customer
        role:
          type: string
          minLength: 1
          description: User's role
          enum:
            - admin
            - cohort-admin
            - founder
            - investor
        hasMarketingConsent:
          type: boolean
          description: User's consent status
        marketingConsentTimestamp:
          type: string
          minLength: 1
          description: Date and time user's consent has been taken
          example: '2020-11-30T17:45:32.193Z'
        watchedVideos:
          type: number
          description: User's count of watched videos
          example: 123
    AggregatedUserProfile:
      allOf:
        - $ref: '#/components/schemas/UserProfileId'
        - $ref: '#/components/schemas/UserProfileUserEditable'
        - $ref: '#/components/schemas/UserProfileAdminEditable'
...
    GenericErrors:
      title: GenericErrors
      type: object
      x-examples:
        example-1:
          - string
      properties:
        errors:
          type: array
          description: Errors list occurred during request processing
          items:
            type: string
      description: ''
...
  securitySchemes:
    API Key:
      name: API Key
      type: apiKey
      in: header
      description: Firebase token